Wording and grammar changes.
I changed some of the pronouns to this reads a little smoother.  I also make sure to spell out what RTO and RPO mean as those terms are not well known outside of the storage and data recovery communities. Otherwise, this is excellent!  Thanks @yasker for all of your hard work.  I love Longhorn